workers/face-worker/requirements.txt
numpy==2.1.3
opencv-python-headless==4.10.0.84
pillow==11.0.0
requests==2.32.3
python-dotenv==1.0.1

onnxruntime==1.20.1
insightface==0.7.3

scikit-learn==1.5.2
hdbscan==0.8.40

supabase==2.10.0


